requirements.txt
pandas>=2.0.0
python-dotenv>=1.0.0
intuit-oauth>=1.2.6
requests>=2.31.0
pytest>=7.4.0
ruff>=0.1.0
